common.pyを共通設定とlog関係に分離
